refactor(bot): update shared types
